tests/wpt/tests/sanitizer-api/sanitizer-unknown.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
test(t => {
  let d = document.createElement("div")
  d.setHTML("<hello><world>",
      { sanitizer: { elements: ["b", "em"] } });
  assert_equals(d.innerHTML, "");
}, "Unknown element names get blocked without being allowed.");

test(t => {
  let d = document.createElement("div")
  d.setHTML("<hello><world>",
      { sanitizer: { elements: ["hello", "world"] } });
  assert_equals(d.innerHTML, "<hello><world></world></hello>");
}, "Unknown element names pass when allowed.");

test(t => {
  let d = document.createElement("div")
  d.setHTML("<b hello='1' world>",
    { sanitizer: { attributes: ["name", "href"] } });
  assert_equals(d.innerHTML, "<b></b>");
}, "Unknown attributes names get blocked without being allowed.");

test(t => {
  let d = document.createElement("div")
  d.setHTML("<b hello='1' world>",
    { sanitizer: { attributes: ["hello", "world"] } });
  assert_equals(d.innerHTML, `<b hello="1" world=""></b>`);
}, "Unknown attribute names pass when allowed.");

test(t => {
  let d = document.createElement("div")
  d.setHTML(`<div is="foo-bar">hello</div>`);
  assert_equals(d.innerHTML, `<div>hello</div>`);
  assert_false(d.firstChild.hasAttribute("is"));
  assert_true(d.firstChild.matches(":defined"));
}, "The is= attribute is removed by the default sanitizer config.");

test(t => {
  let d = document.createElement("div")
  d.setHTML(`<span is="foo-bar">hello</span>`);
  assert_equals(d.innerHTML, `<span>hello</span>`);
  assert_false(d.firstChild.hasAttribute("is"));
  assert_true(d.firstChild.matches(":defined"));
}, "The is= attribute is removed from span by the default sanitizer config.");


test(t => {
  let d = document.createElement("div")
  d.setHTML(`<div is="foo-bar">hello</div>`,
    { sanitizer: { attributes: ["is"] } });
  assert_equals(d.innerHTML, `<div is="foo-bar">hello</div>`);
  assert_true(d.firstChild.hasAttribute("is"));
  assert_false(d.firstChild.matches(":defined"));
}, "The is= attribute is kept when explicitly allowed.");

test(t => {
  let d = document.createElement("div")
  d.setHTML(`<div is="foo-bar">hello</div>`,
    { sanitizer: { attributes: [] } });
  assert_equals(d.innerHTML, `<div>hello</div>`);
  assert_false(d.firstChild.hasAttribute("is"));
  assert_true(d.firstChild.matches(":defined"));
}, "The is= attribute gets blocked with a custom config.");
</script>
</body>
</html>
